streamlit_app.py
import streamlit as st
import requests
import json
import time
from datetime import datetime
import plotly.graph_objects as go
import re
from googlesearch import search as google_search
from youtubesearchpython import VideosSearch
import snscrape.modules.twitter as sntwitter

# Configuração da página
st.set_page_config(page_title="InfluScore Real", page_icon="🔍", layout="wide")

# CSS simplificado
st.markdown("""<style>
body { font-family: 'Inter', sans-serif; }
""", unsafe_allow_html=True)

class RealSearchAnalyzer:
    def __init__(self):
        self.positive_keywords = [...]
        self.negative_keywords = [...]

    def search_web_real(self, query, max_results=10):
        st.info(f"🔍 Buscando no Google por: {query}")
        results = []
        for url in google_search(query, num_results=max_results, lang='pt'):
            results.append({'title': '', 'snippet': '', 'url': url, 'source': 'Google'})
        st.success(f"✅ {len(results)} resultados do Google")
        return results

    def search_youtube_real(self, query, max_results=10):
        st.info(f"📺 Buscando no YouTube por: {query}")
        videosSearch = VideosSearch(query, limit=max_results)
        results = []
        for v in videosSearch.result()['result']:
            results.append({
                'title': v.get('title',''),
                'description': v.get('description',''),
                'url': v.get('link',''),
                'source':'YouTube'
            })
        st.success(f"✅ {len(results)} vídeos do YouTube")
        return results

    def search_twitter_real(self, query, max_results=10):
        st.info(f"🐦 Buscando no Twitter por: {query}")
        tweets = []
        for i, tweet in enumerate(sntwitter.TwitterSearchScraper(query).get_items()):
            if i>=max_results: break
            tweets.append({'text': tweet.content, 'url': f'https://twitter.com/{tweet.user.username}/status/{tweet.id}', 'source':'Twitter'})
        st.success(f"✅ {len(tweets)} tweets do Twitter")
        return tweets

    # ... include analyze_real_content, calculate_real_score, create_real_gauge, main() similar ao anterior ...
