requirements.txt
streamlit
plotly
requests
beautifulsoup4
lxml
googlesearch-python
youtube-search-python
snscrape
